+++ apps/manifest-demo/3012-rspack-js-entry-provider/rspack.config.js
@@ -1,9 +1,10 @@
 // const { registerPluginTSTranspiler } = require('nx/src/utils/nx-plugin.js');
 // registerPluginTSTranspiler();
+const path = require('path');
+const reactPath = path.dirname(require.resolve('react/package.json'));
+const reactDomPath = path.dirname(require.resolve('react-dom/package.json'));
 
 const { composePlugins, withNx, withReact } = require('@nx/rspack');
-
-const path = require('path');
 // const { withModuleFederation } = require('@nx/react/module-federation');
 const {
   ModuleFederationPlugin,
@@ -72,6 +73,17 @@ module.exports = composePlugins(
     // publicPath must be specific url
     config.output.publicPath = 'http://localhost:3012/';
 
+    config.plugins.push({
+      name: 'nx-dev-webpack-plugin',
+      apply(compiler) {
+        compiler.options.devtool = false;
+        compiler.options.resolve.alias = {
+          ...compiler.options.resolve.alias,
+          react: reactPath,
+          'react-dom': reactDomPath,
+        };
+      },
+    });
     config.plugins.push(
       new ModuleFederationPlugin({
         name: 'rspack_js_entry_provider',