+++ packages/manifest/jest.config.js
@@ -16,10 +16,18 @@ module.exports = {
   preset: 'ts-jest',
   transformIgnorePatterns: [
     // Change MODULE_NAME_HERE to your module that isn't being compiled
-    '/node_modules/(?!((@byted/garfish-)|(byted-tea-sdk))).+\\.js$',
+    '/node_modules/',
+    '/dist/',
   ],
   transform: {
-    '^.+\\.(t|j)sx?$': ['@swc/jest'],
+    '^.+\\.(t|j)sx?$': [
+      '@swc/jest',
+      {
+        jsc: {
+          target: 'es2022',
+        },
+      },
+    ],
   },
   rootDir: __dirname,
   testMatch: ['<rootDir>__tests__/**/**.spec.[jt]s?(x)'],